change MainDutyRepository and MainDutyService

+++ src/main/java/sa/aref/repository/duty/MainDutyRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import sa.aref.entity.duties.MainDuties;
 @Repository
-public interface MainDutyRepository extends JpaRepository<MainDuties,Long> {
+public interface MainDutyRepository extends JpaRepository<MainDuties,Integer> {
 }
